webrobotapp/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.6.5
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
Bcrypt-Flask==1.0.2
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
consonance==0.1.5
cryptography==36.0.0
cycler==0.11.0
dissononce==0.34.3
dnspython==2.1.0
dominate==2.6.0
dropbox==11.17.0
email-validator==1.1.3
Flask==1.1.4
flask-blueprint==1.3.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
fonttools==4.28.3
frozenlist==1.2.0
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.2
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.5.0
mediapipe-rpi4==0.8.8
mplayer.py==0.7.2.post1
multidict==5.2.0
numpy==1.21.2
opencv-contrib-python==4.5.3.56
packaging==21.3
pandas==1.3.5
phonenumbers==8.12.39
picamera==1.13
Pillow==8.4.0
pkg-resources==0.0.0
play-mp3==0.0.1
playsound==1.3.0
ply==3.11
protobuf==3.19.1
py-espeak-ng==0.1.8
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pygame==2.1.0
pyparsing==3.0.6
python-axolotl==0.2.2
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.1
python-editor==1.0.4
python-telegram-bot==13.9
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.2.1
readline==6.2.4.1
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.38.1
sqlparse==0.4.1
stone==3.2.1
telepot==12.7
tornado==6.1
transitions==0.8.10
twilio==7.3.2
typing-extensions==3.10.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
yarl==1.7.2
zipp==3.5.0
zmq==0.0.0
